% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Get the standard deviation of a vector}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
The standard deviation of the numeric vector
}
\description{
Get the standard deviation of a vector
}
\examples{
standard_deviation(c(1, 2, 3))
0.8164966
}

