% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error of a vector}
\usage{
standard_error(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
: The standard error
}
\description{
Calculate the standard error of a vector
}
\examples{
standard_error(c(1, 2, 3))
0.4714045
}

