% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standarddeviation-package.r
\docType{package}
\name{standarddeviation}
\alias{standarddeviation}
\alias{standarddeviation-package}
\title{standarddeviation: Do standard deviation}
\description{
If you want a package for calculating the standard deviation
or standard error, you've come to the right place!
}

